+++ test/integration/cartridge_nohttp_test.lua
@@ -4,18 +4,6 @@ local g = t.group('cartridge-without-http')
 
 local helpers = require('test.helper')
 
-local function set_export(cluster, export)
-    local server = cluster.main_server
-    return server.net_box:eval([[
-        local cartridge = require('cartridge')
-        local metrics = cartridge.service_get('metrics')
-        local _, err = pcall(
-            metrics.set_export, ...
-        )
-        return err
-    ]], {export})
-end
-
 g.test_http_disabled = function()
     t.skip_if(type(helpers) ~= 'table', 'Skip cartridge test')
     local cluster = helpers.init_cluster()
@@ -27,13 +15,17 @@ g.test_http_disabled = function()
         cartridge.service_set('httpd', nil)
     ]])
 
-    local ret = set_export(cluster, {
+    local ret = helpers.set_export(cluster, {
         {
             path = '/metrics',
             format = 'json',
         },
     })
     t.assert_not(ret)
+
+    server.net_box:eval([[
+        require('cartridge').service_set('httpd', _G.old_service)
+    ]])
 
     cluster:stop()
     fio.rmtree(cluster.datadir)